FIDALLI-PROJET/requirements.txt
Django>=3.2,<5.0
djangorestframework>=3.12.0,<4.0
djangorestframework-simplejwt>=5.2.0,<6.0
mysqlclient
django-cors-headers
djangorestframework-simplejwt